role fix

+++ admin/realisasi_kuo.php
@@ -70,8 +70,20 @@
 										<td><?=$b['tujuan'] ?></td>
 										<td><?=$b['nopol'] ?></td>
 										<td>
-											<a href="edit_real_kuo.php?id=<?=$b['id']?>&no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
-                      <a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=realisasi_kuo&id=<?=$b['id']?>&no_bukti=<?=$no_bukti?>' }" class="btn btn-danger">Hapus</a>
+                      <?php
+                        if($type=="Admin"){
+                          ?>
+                          <a href="edit_real_kuo.php?id=<?=$b['id']?>&no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
+                          <a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=realisasi_kuo&id=<?=$b['id']?>&no_bukti=<?=$no_bukti?>' }" class="btn btn-danger">Hapus</a>
+                          <?php
+                			  } else if($type=="Editor"){
+                          ?>
+                          <a href="edit_real_kuo.php?id=<?=$b['id']?>&no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
+                          <?php
+                			  } else{
+
+                			  }
+                      ?>
 										</td>
 									</tr>
 									<?php
